+++ src/main.js
@@ -54,7 +54,6 @@ import { isEditing } from './Edit'
 import { CollapseTableButton } from './Detail'
 import { DeleteDetail } from './delete/DeleteDetail'
 import { DeleteButton } from './table/Table'
-import { getInputType } from './form/InputType'
 import { Search } from './Search'
 import { getSearchable } from './utils/schemaGetters'
 
@@ -114,6 +113,5 @@ export { isEditing }
 export { CollapseTableButton }
 export { DeleteDetail }
 export { DeleteButton }
-export { getInputType }
 export { Search }
 export { getSearchable }